Hash rows or columns in a data frame

The function hash_df accepts a data frame and performs an MD5 along the specified margin using the openssl package. When margin = "rows", the original data frame is return with an additional column labeled md5 containing row-level hash values. When margin = "columns", the column names from the original data frame and md5 hash are returned as a data frame.

Usage

hash_df(df, margin = c("rows", "columns"))

Arguments

df

The data frame to be hashed.

margin

A character value indicating the margin on which the hash will be calculated (rows or columns).

Value

A dataframe with two columns: name and hash

Details

The function may be used in the process of determining which columns have changed in a data frame.

Examples

hash_df(iris[1])
#> # A tibble: 150 × 2
#> # Rowwise: 
#>    Sepal.Length md5                             
#>    <chr>        <chr>                           
#>  1 5.1          0b1988431c8b99b756e34790dc4522da
#>  2 4.9          1c6b6cf74bdcf08605f2b032d9c49105
#>  3 4.7          b0f2ea0999e55a634f69471adf767da7
#>  4 4.6          74544dc17009fa1b4d5bc3040ef262f1
#>  5 5            ee340e888492be0703f2bcc9abfb390c
#>  6 5.4          40ab8c9c8ce4885b6f557e0873643ecc
#>  7 4.6          74544dc17009fa1b4d5bc3040ef262f1
#>  8 5            ee340e888492be0703f2bcc9abfb390c
#>  9 4.4          d877ce59a29f544b5a3621d7dfdad2f7
#> 10 4.9          1c6b6cf74bdcf08605f2b032d9c49105
#> # … with 140 more rows
#> # ℹ Use `print(n = ...)` to see more rows
hash_df(iris)
#> # A tibble: 150 × 6
#> # Rowwise: 
#>    Sepal.Length Sepal.Width Petal.Length Petal.Width Species md5                
#>    <chr>        <chr>       <chr>        <chr>       <chr>   <chr>              
#>  1 5.1          3.5         1.4          0.2         setosa  a8c38e2a38db5ec876…
#>  2 4.9          3           1.4          0.2         setosa  23b0260532d7c04122…
#>  3 4.7          3.2         1.3          0.2         setosa  2e4c52b64dd67aa963…
#>  4 4.6          3.1         1.5          0.2         setosa  3282393366d46ec19b…
#>  5 5            3.6         1.4          0.2         setosa  2c0593d3d66905e9fb…
#>  6 5.4          3.9         1.7          0.4         setosa  5402af39b5fcdf28e3…
#>  7 4.6          3.4         1.4          0.3         setosa  971ff22101798bc2f5…
#>  8 5            3.4         1.5          0.2         setosa  53ddb7a21421b98a83…
#>  9 4.4          2.9         1.4          0.2         setosa  e5cb67a6faef6748ee…
#> 10 4.9          3.1         1.5          0.1         setosa  aeb834abb3707995dd…
#> # … with 140 more rows
#> # ℹ Use `print(n = ...)` to see more rows